refactor(clickhouse): split reader module
