+++ crates/clickhouse/src/reader/mod.rs
@@ -1,0 +1,8 @@
+mod client;
+mod time_range;
+
+pub use client::ClickhouseReader;
+pub use time_range::TimeRange;
+
+#[cfg(test)]
+mod tests;
